+++ app/activitypub_utils.py
@@ -18,7 +18,9 @@ from datetime import datetime, timezone
 from email.utils import formatdate
 from urllib.parse import urlparse
 from flask import Blueprint, current_app, request, abort, jsonify, url_for
-from app.models import User, ActivityStore, QuestLike, db, Notification, QuestSubmission
+from app.models import db
+from app.models.user import User, ActivityStore, Notification
+from app.models.quest import QuestLike, QuestSubmission
 
                                      
 ap_bp = Blueprint('activitypub', __name__)
@@ -297,7 +299,7 @@ def inbox(username):
             if '/submissions/' in in_to:
                 sid = int(in_to.rsplit('/',1)[1])
                                                 
-                from app.models import SubmissionReply
+                from app.models.quest import SubmissionReply
                 reply = SubmissionReply(
                     submission_id=sid,
                     user_id=(sender.id if sender else None),
