+++ app/models/notification.py
@@ -1,0 +1,1 @@
+from .user import Notification
