+++ tests/dumpPrivKey.py
@@ -1,5 +1,5 @@
 from app import create_app
-from app.models import User
+from app.models.user import User
 app = create_app()
 with app.app_context():
     user = User.query.filter_by(username='t12').first()
