+++ tests/test_activitypub_delivery.py
@@ -2,7 +2,7 @@ import pytest
 from unittest.mock import patch
 
 from app import create_app, db
-from app.models import User
+from app.models.user import User
 from app.activitypub_utils import (
     deliver_activity,
     generate_activitypub_keys,
