+++ tests/test_auth.py
@@ -3,7 +3,7 @@ import pytest
 from urllib.parse import urlparse, parse_qs
 
 from app import create_app, db
-from app.models import User
+from app.models.user import User
 from flask import url_for
 from datetime import datetime, timezone
