+++ tests/test_auth_extra.py
@@ -3,7 +3,7 @@
 import pytest
 from flask import url_for
 from app import create_app, db
-from app.models import User
+from app.models.user import User
 from datetime import datetime, timezone
 
 @pytest.fixture
