+++ tests/test_registration.py
@@ -2,7 +2,7 @@ import pytest
 from urllib.parse import urlparse, parse_qs
 
 from app import create_app, db
-from app.models import User
+from app.models.user import User
 from flask import url_for
 from datetime import datetime, timezone
 from sqlalchemy.exc import SQLAlchemyError
